requirements.txt
anyio==3.6.1
cairocffi==1.5.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
coloredlogs==15.0.1
contourpy==1.0.7
cryptography==36.0.2
cssselect2==0.7.0
cycler==0.11.0
defusedxml==0.7.1
falcon==3.1.0
fastapi==0.79.0
filelock==3.12.0
fonttools==4.39.4
fsspec==2023.5.0
grpc-reflection==0.0.1
grpcio==1.54.2
grpcio-reflection==1.54.2
grpcio-tools==1.54.2
gunicorn==20.1.0
h11==0.13.0
h5py==3.8.0
html5lib==1.1
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.3
importlib-resources==5.12.0
Jinja2==3.1.2
jose==1.0.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
Pillow==9.5.0
protobuf==4.23.1
pyclipper==1.3.0.post4
pycparser==2.21
pydantic==1.9.1
PyJWT==2.3.0
pyparsing==3.0.9
PyPDF2==2.10.2
pypdfium2==2.11.0
pyphen==0.14.0
python-dateutil==2.8.2
PyYAML==6.0
rapidfuzz==3.0.0
reflection==0.1.1
requests==2.27.1
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
sympy==1.12
tinycss2==1.2.1
torch==2.0.1
tqdm==4.65.0
typing_extensions==4.3.0
Unidecode==1.3.6
urllib3==1.26.9
uvicorn==0.18.2
webencodings==0.5.1
zipp==3.15.0
